Fast Food Restaurant
It is my individual work. There is a fast food section, a drink and a dessert section in the project. There are 6 types in total in both sections. It includes the selected varieties in the menu and calculates the price. There is also a part for deleting and printing receipts. The program is all in English source code.